feat: スタッフ画面のレスポンシブデザイン実装完了
## 主な変更点

### レスポンシブデザイン対応
- スタッフ側全画面にモバイル対応を実装
- ハンバーガーメニューとモバイルヘッダーの追加
- タブレット・モバイルでの表示最適化

### 新規コンポーネント
- MobileHeader: モバイル用ヘッダー
- _responsive-staff.scss: スタッフ側レスポンシブスタイル

### ダッシュボード改善
- 新しいオファーセクションの追加と表示修正
- 給与明細テーブルのモバイル最適化（列の調整、フォントサイズ縮小）
- カード形式とテーブル形式の自動切り替え

### マイページ改善
- 評価一覧のレスポンシブ対応
- PC: テーブル表示
- モバイル: カード形式表示
- ID列の削除、訪問日を最上部に移動
- 実施事業所の重複表示を削除

### その他の改善
- サイドバーのモバイル対応（テキストのみ表示、アイコン非表示）
- タブレット画面での初期状態修正（サイドバー非表示）
- ヘッダーとコンテンツの重なり問題を修正

### デュアル予約システム
- 社員自己予約機能の実装
- 企業側予約との統合
- データ整合性の修正と改善
- TimeSlotDisplay コンポーネントの判定ロジック改善

### ドキュメント追加
- STAFF_RESPONSIVE_DESIGN.md
- DUAL_BOOKING_IMPLEMENTATION_COMPLETE.md
- DATA_CONSISTENCY_FIX.md
- EMPLOYEE_ID_DISPLAY_FIX_SUMMARY.md
- PROGRESS_SUMMARY.md
- TEST_DATA_012_013_GUIDE.md

+++ backend/app/schemas/staff.py
@@ -16,7 +16,7 @@ class StaffBase(BaseModel):
     available_days: Optional[str] = None
     line_id: Optional[str] = None
     is_available: bool = True
-    rating: Optional[int] = None
+    rating: Optional[float] = None  # float型に修正（平均評価は小数点を含む）
     notes: Optional[str] = None
